Fixing issues with uploading images.
There was an issue with file directory linking, on the action of uploading a new image to a project. The issue was resolved by root linking all of the filepaths ie."/var/www/path/to" previously used was "./". The previous solution  failed after launching into alpha. Still needs to be diagnosed, however the current implimentation is solid.